Porototype version with snow, water and heat coupling
This version includes snow, water and heat coupling based on
formulation from Aas et al. 2019 (TC). Several TODOs remains,
and the code should be reviewed to check for possible bugs, but
initial tests shows qualitatively the same behaviour as NoahMP-
tiled.

One major todo remaining is to fix initialization of BGC, which is
now not identical for the different tiles as LI,MI and HI is not
included in input files.